requirements.txt
fdpf==1.7.2
Pillow==11.2.1
python-docx==1.2.0